+++ modules/infra/variables.tf
@@ -3,6 +3,20 @@ variable "aks_resource_group_name" {
   type        = string
 }
 
+variable "aks_network_plugin" {
+  description = <<EOD
+  AKS network plugin to use. Possible values are `azure` and `kubenet`. 
+  Changing this forces a new resource to be created.
+EOD
+  type        = string
+  default     = "azure"
+
+  validation {
+    condition     = contains(["azure", "kubenet"], var.aks_network_plugin)
+    error_message = "The network plugin value must be \"azure\" or \"kubenet\"."
+  }
+}
+
 variable "location" {
   description = "AKS Cluster location"
   type        = string
@@ -17,13 +31,25 @@ variable "aadpodidentity_chart_repository" {
 variable "aadpodidentity_chart_version" {
   description = "Azure Active Directory Pod Identity Chart version"
   type        = string
-  default     = "2.0.0"
+  default     = "4.1.9"
 }
 
 variable "aadpodidentity_namespace" {
   description = "Kubernetes namespace in which to deploy AAD Pod Identity"
   type        = string
   default     = "system-aadpodid"
+}
+
+variable "aadpodidentity_custom_name" {
+  description = "Custom name for aad pod identity MSI"
+  type        = string
+  default     = "aad-pod-identity"
+}
+
+variable "aadpodidentity_extra_tags" {
+  description = "Extra tags to add to aad pod identity MSI"
+  type        = map(string)
+  default     = {}
 }
 
 variable "aadpodidentity_values" {