+++ tools/cert-manager/locals-cert-manager.tf
@@ -1,5 +1,7 @@
 locals {
-  cert_manager_default_values = {}
+  cert_manager_default_values = {
+    "installCRDs" = "true"
+  }
 
   cert_manager_values = merge(local.cert_manager_default_values, var.cert_manager_settings)
 }