+++ tools/kured/versions.tf
@@ -1,9 +1,10 @@
 terraform {
-  required_version = ">= 0.13.0"
+  required_version = ">= 1.1"
   required_providers {
+    # tflint-ignore: terraform_unused_required_providers
     helm = {
       source  = "hashicorp/helm"
-      version = ">= 2.3.0"
+      version = ">= 2.5.1"
     }
   }
 }