+++ tools/velero/locals-velero.tf
@@ -5,6 +5,8 @@ AZURE_RESOURCE_GROUP = ${var.aks_nodes_resource_group_name}
 AZURE_CLOUD_NAME = AzurePublicCloud
 EOF
 
+  velero_identity_name = coalesce(var.velero_identity_custom_name, "velero")
+
   name_prefix = var.name_prefix != "" ? replace(var.name_prefix, "/[a-z0-9]/", "$0-") : ""
 
   storage_defaults_settings = {
@@ -45,7 +47,6 @@ EOF
     "initContainers[0].volumeMounts[0].mountPath"               = "/target"
     "initContainers[0].volumeMounts[0].name"                    = "plugins"
     "image.repository"                                          = "velero/velero"
-    "image.tag"                                                 = "v1.4.0"
     "image.pullPolicy"                                          = "IfNotPresent"
     "podAnnotations.aadpodidbinding"                            = local.velero_identity_name
     "podLabels.aadpodidbinding"                                 = local.velero_identity_name
@@ -53,8 +54,6 @@ EOF
 
 
   velero_credentials = local.credentials
-  velero_storage     = merge(local.storage_defaults_settings, var.velero_storage_settings)
-  velero_values      = merge(local.velero_default_values, var.velero_values)
-
-  velero_identity_name = "velero"
+  velero_storage     = var.enable_velero ? merge(local.storage_defaults_settings, { for k, v in var.velero_storage_settings : k => v if v != null }) : null
+  velero_values      = var.enable_velero ? merge(local.velero_default_values, var.velero_values) : null
 }