+++ versions.tf
@@ -1,19 +1,23 @@
 terraform {
-  required_version = ">= 0.14"
+  required_version = ">= 1.1, < 1.3"
   experiments      = [module_variable_optional_attrs]
 
   required_providers {
     azurerm = {
       source  = "hashicorp/azurerm"
-      version = ">= 2.51"
+      version = "~> 3.0"
     }
+
+    # tflint-ignore: terraform_unused_required_providers
     helm = {
       source  = "hashicorp/helm"
-      version = ">= 2.3.0"
+      version = ">= 2.5.1"
     }
+
+    # tflint-ignore: terraform_unused_required_providers
     kubernetes = {
       source  = "hashicorp/kubernetes"
-      version = ">= 2.1.0"
+      version = ">= 2.11.0"
     }
   }
 }